requirements.txt
# CDK Dependencies (for development)
aws-cdk-lib>=2.150.0
constructs>=10.0.0
python-dotenv>=1.0.0

# Lambda Dependencies (for runtime)
requests>=2.31.0
boto3>=1.34.0